+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/notification/domain/base/service/AbstractNotificationReceiverService.java
@@ -44,11 +44,13 @@ public abstract class AbstractNotificationReceiverService implements Notificatio
 
     protected abstract RuntimeException getIllegalUpdateException(String message);
 
+    protected abstract BPN getApplicationBpn();
+
     @Override
     public void handleReceive(EDCNotification edcNotification, NotificationType notificationType) {
         BPN investigationCreatorBPN = BPN.of(edcNotification.getSenderBPN());
         NotificationMessage notification = getNotificationMessageMapper().toNotificationMessage(edcNotification, notificationType);
-        Notification investigation = getNotificationMapper().toNotification(investigationCreatorBPN, edcNotification, notification, notificationType);
+        Notification investigation = getNotificationMapper().toNotification(investigationCreatorBPN, edcNotification, notification, notificationType, getApplicationBpn());
         NotificationId investigationId = getRepository().saveNotification(investigation);
         log.info("Stored received edcNotification in investigation with id {}", investigationId);
     }