+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/notification/application/alert/response/AlertResponseTest.java
@@ -61,7 +61,7 @@ class AlertResponseTest {
                 .hasFieldOrPropertyWithValue("createdDate", notification.getCreatedAt().toString())
                 .hasFieldOrPropertyWithValue("assetIds", notification.getAffectedPartIds())
                 .hasFieldOrPropertyWithValue("channel", NotificationSideResponse.SENDER)
-                .hasFieldOrPropertyWithValue("sendTo", "recipientBPN")
+                .hasFieldOrPropertyWithValue("sendTo", "TESTBPN")
                 .hasFieldOrPropertyWithValue("sendToName", "receiverManufacturerName")
                 .hasFieldOrPropertyWithValue("severity", NotificationSeverityResponse.MINOR)
                 .hasFieldOrPropertyWithValue("targetDate", null);