+++ CHANGELOG.md
@@ -10,6 +10,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Changed
 - #965 Implement proxy functionality of the IRS policy store
 - #962 Changed notification model to new one in frontend/backend
+- #962 Removed initial notification message for notification flow
 - #753 Refactored message history in notification detail view
 - XXX updated local deployment documentation